refactor(backend, shared): fix backend/shared separation and compilation
