+++ packages/backend/src/controllers/private-rest.unauthorized.spec.ts
@@ -9,7 +9,7 @@ import {
 import request from "supertest";
 import express from "express";
 import { createExpressServer } from "@createExpressServer";
-import { isInDebugMode } from "@shared/util/debug";
+import { isInDebugMode } from "@util/debug";
 import { CreateProjectProps } from "@shared/domain/writeModels/project/WriteProject";
 import { decodeJwt } from "jose";
